+++ packages/tswebextension/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Updated [@adguard/agtree] to `v4.0.2`.
+<!-- TODO: specify the version later -->
+- Updated [@adguard/tsurlfilter] to `v4.0.X`.
+
 ## [4.0.3] - 2026-02-26
 
 ### Added
@@ -88,10 +96,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Memory leak caused by storing data URL favicons in tab contexts. [AdguardBrowserExtension#2594]
+- Memory leak caused by storing data URL favicons in tab contexts [AdguardBrowserExtension#2594].
 
 [3.2.19]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.2.19
-[AdguardBrowserExtension#2594]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2594
 
 ## [3.2.18] - 2025-12-16
 
@@ -471,7 +478,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remade JS rules injections in MV3:
     - use `chrome.scripting` API for injecting functions for script rules from the pre-built filters,
     - use script tag injection only for script rules manually added by users —
-      rules from *User rules* and *Custom filters*.
+      rules from _User rules_ and _Custom filters_.
 
 [2.4.0-alpha.8]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v2.4.0-alpha.8