+++ packages/tswebextension/CHANGELOG.md
@@ -5,13 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [4.0.4] - 2026-03-18
 
 ### Changed
 
 - Updated [@adguard/agtree] to `v4.0.2`.
-<!-- TODO: specify the version later -->
-- Updated [@adguard/tsurlfilter] to `v4.0.X`.
+- Updated [@adguard/tsurlfilter] to `v4.0.3`.
+- Updated [@adguard/scriptlets] to `v2.3.0`.
+
+[4.0.4]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v4.0.4
 
 ## [4.0.3] - 2026-02-26